slackchatbot/requirements.txt
slackclient==2.8.1
PyYAML==5.3.1
nltk==3.5
scikit-learn==0.23.2
